refactor(swarm-ui): fold seed-phrase helpers into crypto/mnemonic
`seed-phrase.ts` was a standalone file holding two BIP-39 phrase helpers
(`validateSeedPhrase`, `countSeedPhraseWords`) whose normalize + parse logic
overlapped `crypto/mnemonic.ts`. Move them there — co-located with the rest of
the phrase/mnemonic code, mirroring where `ui/` keeps `isValidPhrase` — update
the two phrase-entry importers (agent/new, import/passkey), and delete the
file. No behavior change.
